Move node utils
