+++ src/utils/nodes/node.ts
@@ -7,8 +7,8 @@
 import { Attrs, Node, ResolvedPos, TagParseRule } from "@tiptap/pm/model";
 import { Transaction } from "@tiptap/pm/state";
 import { Attributes } from "@tiptap/core";
-import { NodeAttributes } from "../types/node";
-import { wrapJSONParse } from "./object";
+import { NodeAttributes } from "../../types/node";
+import { wrapJSONParse } from "../object";
 
 /**
  * Get the type of the node at the specified position.
