+++ src/utils/paragraph.ts
@@ -6,7 +6,7 @@
 
 import { Node as PMNode, ResolvedPos } from "@tiptap/pm/model";
 import { Nullable } from "../types/record";
-import { getParentNodePosOfType, isNodeEmpty } from "./node";
+import { getParentNodePosOfType, isNodeEmpty } from "./nodes/node";
 import { isPosAtEndOfDocument, isPosAtStartOfDocument } from "./pagination";
 import { inRange } from "./math";
